Proyecto 2/requirements.txt
cmake
dlib
face_recognition
numpy
opencv-python
mysql-connector